changed structure, added image rescaling
