+++ requirements.txt
@@ -1,70 +1,70 @@
 # This file may be used to create an environment using:
 # $ conda create --name <env> --file <this file>
 # platform: linux-64
-blas=1.0=mkl
-bzip2=1.0.6=h14c3975_5
-ca-certificates=2018.03.07=0
-cairo=1.14.12=h8948797_3
-certifi=2018.11.29=py37_0
-cycler=0.10.0=py37_0
-dbus=1.13.2=h714fa37_1
-expat=2.2.6=he6710b0_0
-ffmpeg=4.0=hcdf2ecd_0
-fontconfig=2.13.0=h9420a91_0
-freeglut=3.0.0=hf484d3e_5
-freetype=2.9.1=h8a8886c_1
-glib=2.56.2=hd408876_0
-graphite2=1.3.12=h23475e2_2
-gst-plugins-base=1.14.0=hbbd80ab_1
-gstreamer=1.14.0=hb453b48_1
-harfbuzz=1.8.8=hffaf4a1_0
-hdf5=1.10.2=hba1933b_1
-icu=58.2=h9c2bf20_1
-intel-openmp=2019.1=144
-jasper=2.0.14=h07fcdf6_1
-jpeg=9b=h024ee3a_2
-kiwisolver=1.0.1=py37hf484d3e_0
-libedit=3.1.20170329=h6b74fdf_2
-libffi=3.2.1=hd88cf55_4
-libgcc-ng=8.2.0=hdf63c60_1
-libgfortran-ng=7.3.0=hdf63c60_0
-libglu=9.0.0=hf484d3e_1
-libopencv=3.4.2=hb342d67_1
-libopus=1.3=h7b6447c_0
-libpng=1.6.35=hbc83047_0
-libstdcxx-ng=8.2.0=hdf63c60_1
-libtiff=4.0.9=he85c1e1_2
-libuuid=1.0.3=h1bed415_2
-libvpx=1.7.0=h439df22_0
-libxcb=1.13=h1bed415_1
-libxml2=2.9.8=h26e45fe_1
-matplotlib=3.0.2=py37h5429711_0
-mkl=2019.1=144
-mkl_fft=1.0.6=py37hd81dba3_0
-mkl_random=1.0.2=py37hd81dba3_0
-ncurses=6.1=he6710b0_1
-numpy=1.15.4=py37h7e9f1db_0
-numpy-base=1.15.4=py37hde5b4d6_0
-opencv=3.4.2=py37h6fd60c2_1
-openssl=1.1.1a=h7b6447c_0
-pcre=8.42=h439df22_0
-pip=18.1=py37_0
-pixman=0.34.0=hceecf20_3
-py-opencv=3.4.2=py37hb342d67_1
-pyparsing=2.3.0=py37_0
-pyqt=5.9.2=py37h05f1152_2
-python=3.7.1=h0371630_7
-python-dateutil=2.7.5=py37_0
-pytz=2018.7=py37_0
-qt=5.9.7=h5867ecd_1
-readline=7.0=h7b6447c_5
-scipy=1.1.0=py37h7c811a0_2
-setuptools=40.6.3=py37_0
-sip=4.19.8=py37hf484d3e_0
-six=1.12.0=py37_0
-sqlite=3.26.0=h7b6447c_0
-tk=8.6.8=hbc83047_0
-tornado=5.1.1=py37h7b6447c_0
-wheel=0.32.3=py37_0
-xz=5.2.4=h14c3975_4
-zlib=1.2.11=h7b6447c_3
+# blas # DE
+# bzip2  # DE
+# ca-certificates==2018.03.07
+pycairo==1.14.12
+certifi==2018.11.29
+cycler==0.10.0
+dbus==1.13.2
+expat==2.2.6
+ffmpeg==4.0
+fontconfig==2.13.0
+freeglut==3.0.0
+freetype==2.9.1
+glib==2.56.2
+graphite2==1.3.12
+gst-plugins-base==1.14.0
+gstreamer==1.14.0
+harfbuzz==1.8.8
+hdf5==1.10.2
+icu==58.2
+intel-openmp==2019.1
+jasper==2.0.14
+jpeg==9b
+kiwisolver==1.0.1
+libedit==3.1.20170329
+libffi==3.2.1
+libgcc-ng==8.2.0
+libgfortran-ng==7.3.0
+libglu==9.0.0
+libopencv==3.4.2
+libopus==1.3
+libpng==1.6.35
+libstdcxx-ng==8.2.0
+libtiff==4.0.9
+libuuid==1.0.3
+libvpx==1.7.0
+libxcb==1.13
+libxml2==2.9.8
+matplotlib==3.0.2
+mkl==2019.1
+mkl_fft==1.0.6
+mkl_random==1.0.2
+ncurses==6.1
+numpy==1.15.4
+numpy-base==1.15.4
+opencv==3.4.2
+openssl==1.1.1a
+pcre==8.42
+pip==18.1
+pixman==0.34.0
+py-opencv==3.4.2
+pyparsing==2.3.0
+pyqt==5.9.2
+python==3.7.1
+python-dateutil==2.7.5
+pytz==2018.7
+qt==5.9.7
+readline==7.0
+scipy==1.1.0
+setuptools==40.6.3
+sip==4.19.8
+six==1.12.0
+sqlite==3.26.0
+tk==8.6.8
+tornado==5.1.1
+wheel==0.32.3
+xz==5.2.4
+zlib==1.2.11
